# todoswagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "API for managing todos",
    "version" : "2023-11-28T11:18:05Z",
    "title" : "dev-aws-node-express-api-project"
  },
  "host" : "a4ktsp0ijd.execute-api.eu-central-1.amazonaws.com",
  "basePath" : "/dev",
  "schemes" : [ "https" ],
  "paths" : {
...
    "/deleteTodo/{id}" : {
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : { },
        "x-amazon-apigateway-integration" : {
          "httpMethod" : "POST",
          "uri" : "arn:aws:apigateway:eu-central-1:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-central-1:506820257931:function:aws-node-express-api-project-dev-deleteTodo/invocations",
          "passthroughBehavior" : "when_no_match",
          "type" : "aws_proxy"
        }
      }
    },
...
  },
...
}